+++ tests/fields/test_float_fields.py
@@ -2,6 +2,7 @@ import pytest
 from invana_py.ogm.exceptions import FieldValidationError
 from invana_py.ogm.fields import StringProperty, FloatProperty
 from invana_py.ogm.models import VertexModel
+from gremlin_python.statics import FloatType
 from datetime import datetime
 from invana_py import InvanaGraph
 
@@ -26,13 +27,13 @@ class TestFloatField:
 
     def test_field(self):
         graph.g.V().drop()
-        star = Star.objects.create(name="Sun", distance_from_earth=1989000000000000000000000000000)
+        star = Star.objects.create(name="Sun", distance_from_earth=FloatType(1989000000000000000000000000000))
         assert isinstance(star.properties.distance_from_earth, float)
 
     def test_field_max_value(self):
         graph.g.V().drop()
         with pytest.raises(FieldValidationError) as exec_info:
-            Star.objects.create(name="Sun", distance_from_earth=1989000000000000000000000000000 * 10000)
+            Star.objects.create(name="Sun", distance_from_earth=FloatType(1989000000000000000000000000000 * 10000))
         assert "max_value for field" in exec_info.value.__str__()
 
     def test_field_min_value(self):
